update delete function

+++ src/app/api/posts/myneed/[id]/route.js
@@ -1,14 +1,29 @@
 import { NextResponse } from "next/server";
+import { getAuthSession } from "../../../../../utils/auth";
 
 export const DELETE = async (req, { params }) => {
+  const session = await getAuthSession();
   const { id } = params;
-  await prisma.post.delete({
-    where: {
-      id: id,
-    },
-  });
-  return new NextResponse(
-    JSON.stringify({ message: "Post deleted successfully" }),
-    { status: 202 }
-  );
+  if (!session) {
+    return new NextResponse(
+      JSON.stringify({ message: "Not Authenticated!" }, { status: 401 })
+    );
+  }
+
+  try {
+    await prisma.post.delete({
+      where: {
+        id: id,
+      },
+    });
+    return new NextResponse(
+      JSON.stringify({ message: "Post deleted successfully" }),
+      { status: 204 }
+    );
+  } catch (error) {
+    console.log(error);
+    return new NextResponse(
+      JSON.stringify({ message: "Something went wrong" }, { status: 500 })
+    );
+  }
 };
